requirements.txt
selenium==4.26.1
pytest==8.3.3
allure-pytest==2.13.5